fix: comment, post 삭제 메서드 변경

+++ src/main/java/com/elice/tripnote/domain/post/service/PostService.java
@@ -283,7 +283,9 @@ public class PostService {
 
         Post post = postOrElseThrowsException(postId);
 
-        commentService.deleteCommentsByPostId(postId);
+        if(!post.isDeleted()){
+            commentService.deleteCommentsByPostId(postId);
+        }
 
         post.delete();
